Put client/url stuff in standalone file
